requirements.txt
Flask==3.0.3
PyMuPDF==1.24.1
Pillow==10.3.0